UpdateAdminView

+++ Wagebat/Controllers/CategoriesController.cs
@@ -137,14 +137,17 @@ namespace Wagebat.Controllers
         }
 
         // POST: Categories/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        [HttpPost]
+        [Authorize(Roles = "admin")]
+        public async Task<JsonResult> DeleteConfirmed(int id)
         {
             var category = await _context.Categories.FindAsync(id);
+            if (category == null)
+                return Json(false);
+
             _context.Categories.Remove(category);
             await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            return Json(true);
         }
 
         private bool CategoryExists(int id)
